// apps/web-svelte/src/lib/api/generated/model/libraryBrowseResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { LibraryBrowseEntry } from './libraryBrowseEntry';

export interface LibraryBrowseResponse {
  path: string;
  /** @nullable */
  parentPath: string | null;
  directories: LibraryBrowseEntry[];
}