// apps/web-svelte/src/lib/api/generated/model/settingsGroup.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { SettingDescriptor } from './settingDescriptor';

export interface SettingsGroup {
  key: string;
  label: string;
  description: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  order: number | string;
  settings: SettingDescriptor[];
}